// emp-mgm/src/app/model/roles.model.ts
export class Roles {
    _id?: string;
    name: string;
    permission: Rolespermission;
    editable: boolean;
}

export class Rolespermission {
// ...
    notifications: {
        send?: boolean;
        delete?: boolean;
        view?: boolean;
        approval?: boolean;
    };
// ...
    referrals: {
        viewreferrals?: boolean;
        viewtransactions?: boolean;
        addtransactions?: boolean;
        changestatus?: boolean;
        export?: boolean;
        settings?: boolean
    };
    devices: {
        viewdevices?: boolean;
        export?: boolean;
        settings?: boolean;
        viewloginhistory?: boolean;
        logoutusers?: boolean;
    };
    supports: {
        liveMarketSupportChat?: boolean
        settings?: boolean
        liveMarketSupportReport?: boolean
        export?: boolean;
    };
// ...
    invesmentor: {
        approve?: boolean;
        add?: boolean;
        update?: boolean;
        view?: boolean;
        delete?: boolean;
        viewregistrations?: boolean;
    };
    reports: {
        bcmbreports?: boolean;
        ownbcmbreportins?: boolean;
        ivrreports?: boolean;
        ownivrreports?: boolean;
        owndeptivrreports?: boolean;
        ownteamivrreports?: boolean;
        appreports?: boolean;
    };
    attendanceleave: {
        approveAttendance?: boolean;
        exportAttendance?: boolean;
        viewAllAttendance?: boolean;
        viewTeamAttendance?: boolean;
        editAttendance?: boolean;
        approveRegularize?: boolean;
        exportRegularize?: boolean;
        viewAllRegularize?: boolean;
        viewTeamRegularize?: boolean;
        editRegularize?: boolean;
        deleteRegularize?: boolean;
        approveLeave?: boolean;
        exportLeave?: boolean;
        viewAllLeave?: boolean;
        viewTeamLeave?: boolean;
        deleteLeave?: boolean;
        addLeaveBalance?: boolean;
        manageLeavePolicy?: boolean;
        manageAttendanceSettings?: boolean;
        viewownteamroster?: boolean;
        viewallroster?: boolean;
        viewowndeptroster?: boolean;
        createroster?: boolean;
        deleteroster?: boolean;
        approveroster?: boolean;
    };
    orders: {
        viewAllOrder?: boolean;
        viewOwnOrder?: boolean;
        viewTeamOrder?: boolean;
        createOrder?: boolean;
        updateOrder?: boolean;
        modifyInstallment?: boolean;
        verifyInstallment?: boolean;
        requestAdjustment?: boolean;
        approveAdjustment?: boolean;
        shortClose?: boolean;
        requestRefund?: boolean;
        viewRefundList?: boolean;
        approveRefund?: boolean;
        approveBatchChange?: boolean;
        viewBatchChangeList?: boolean;
        requestBatchChange?: boolean;
        viewPaymentHistory?: boolean;
        changeCourseExpiry?: boolean;
        export?: boolean;
        addPayments?: boolean;
        cancelPaymentLink?: boolean;
    }
}